1. Solve Naming conventions. 2. Enhance test cases, factory, method, classes, interface. 3. Almost done payslip calculator logic based on testing result.

+++ CodingExercise.IncomeTaxCalculator/Interface/ITaxIncomeCalculator.cs
@@ -11,6 +11,6 @@ namespace CodingExercise.IncomeTaxCalculator.Interface
     /// </summary>
     public interface ITaxIncomeCalculator
     {
-        double CalculateIncomeTax(int annualSalary);
+        int CalculateIncomeTax(int annualSalary);
     }
 }
